chore: Remove unused helpers
